Move use-auth hooks
