+++ src/pages/boards/[id].tsx
@@ -9,7 +9,7 @@ import { ColumnList } from "src/features/column/column-list";
 import { Subnav } from "src/features/board/subnav";
 import { imagePath } from "src/utils/image-paths";
 import { getBoardThunk, selectBoardById } from "src/features/board/board.slice";
-import { useAuth } from "src/utils/use-auth";
+import { useAuth } from "src/features/auth/use-auth";
 import { path } from "src/utils/url/drello-web";
 
 const Main = styled.main`
